Begin forcing out the steam run-time.
Compilation works completely; linking does not. Looks like only libstdc++ failes to link now.
i386 libs are searched for in /usr/lib32 only right now. If that is not the case for you, search for that string in src/devtools/makefile_base_posix.mak
Warnings disabled in this patch, temporarily.